+++ src/scene/Mesh.cpp
@@ -1,13 +1,13 @@
 #include "Mesh.hpp"
 
-Mesh::Mesh(const VulkanDevice& device, const std::string& name, const std::vector<Vertex>& vertices, const std::vector<uint16_t>& indices)
+Mesh::Mesh(const VulkanDevice& device, const std::string& name, const std::vector<Vertex>& vertices, const std::vector<uint32_t>& indices)
 	: m_Name(name)
 {
     m_VertexCount = static_cast<uint32_t>(vertices.size());
     m_IndexCount = static_cast<uint32_t>(indices.size());
 
     vk::DeviceSize vertexBufferSize = sizeof(Vertex) * vertices.size();
-    vk::DeviceSize indexBufferSize = sizeof(uint16_t) * indices.size();
+    vk::DeviceSize indexBufferSize = sizeof(uint32_t) * indices.size();
 
     VulkanBuffer vertexStagingBuffer(
         device,
